pyproject.toml
[tool.poetry]
name = "tg-chatbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.12.0"
pydantic = {extras=["email","settings"],version="^2.4.2"}
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
fastapi = "^0.103.2"
passlib = "^1.7.4"
emails = "^0.6"
sqlalchemy = "^2.0.22"
uvicorn = "^0.23.2"
python-jose = {extras = ["cryptography"], version = "^3.1.0"}
bcrypt = "^4.0.1"
python-multipart = "^0.0.6"
pytest = "^7.4.2"
httpx = "^0.25.0"
black = "^23.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
